Create tributepage.html
My second certification project on freecodecamp - the Tribute Page, which is a part of the responsive web design. This project help build a fully functional and responsive tribute webpage using only HTML and CSS. This helps in understanding how the structure and code behind the scenes impacts both user experience and browser behavior.